+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.8.0]
+
+### Added
+
+- Arcjet for bot detection, rate limiting, and WAF protection
+
+### Changed
+
+- UI to be streamlined and include thoughts and tool calls in line with the
+  conversation
+- `freeMessage` default is now 10 messages
+
 ## [0.7.1]
 
 ### Added